requirements.txt
faiss-cpu
requests
tqdm
PyPDF2
python-docx
numpy
pandas
openpyxl
python-pptx
xlrd 